requirements.txt
asgiref==3.5.0
Django==3.2.12
django-jalali-date==1.0.1
jdatetime==4.1.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
